requirements.txt

fastapi
uvicorn
playwright
google-genai
pillow==11.3.0
streamlit
requests
python-dotenv
